VAECache: skip any previously processed files

+++ helpers/vae_cache.py
@@ -3,8 +3,9 @@ from tqdm import tqdm
 from PIL import Image
 import torchvision.transforms as transforms
 
-logger = logging.getLogger('VAECache')
-logger.setLevel('DEBUG')
+logger = logging.getLogger("VAECache")
+logger.setLevel("DEBUG")
+
 
 class VAECache:
     def __init__(self, vae, accelerator, cache_dir="vae_cache", resolution: int = 1024):
@@ -35,17 +36,27 @@ class VAECache:
         filename = os.path.join(self.cache_dir, file_hash + ".pt")
         if os.path.exists(filename):
             latents = self.load_from_cache(filename)
-            logger.debug(f'Loading latents of shape {latents.shape} from existing cache file: {filename}')
+            logger.debug(
+                f"Loading latents of shape {latents.shape} from existing cache file: {filename}"
+            )
         else:
             with torch.no_grad():
-                latents = self.vae.encode(pixel_values.unsqueeze(0).to(self.accelerator.device, dtype=torch.bfloat16)).latent_dist.sample()
-                logger.debug(f'Using shape {latents.shape}, creating new latent cache: {filename}')
+                latents = self.vae.encode(
+                    pixel_values.unsqueeze(0).to(
+                        self.accelerator.device, dtype=torch.bfloat16
+                    )
+                ).latent_dist.sample()
+                logger.debug(
+                    f"Using shape {latents.shape}, creating new latent cache: {filename}"
+                )
             latents = latents * self.vae.config.scaling_factor
-            logger.debug(f'Latent shape after re-scale: {latents.shape}')
+            logger.debug(f"Latent shape after re-scale: {latents.shape}")
             self.save_to_cache(filename, latents.squeeze())
 
-        output_latents = latents.squeeze().to(self.accelerator.device, dtype=self.vae.dtype)
-        logger.debug(f'Output latents shape: {output_latents.shape}')
+        output_latents = latents.squeeze().to(
+            self.accelerator.device, dtype=self.vae.dtype
+        )
+        logger.debug(f"Output latents shape: {output_latents.shape}")
         return output_latents
 
     def process_directory(self, directory):
@@ -54,39 +65,51 @@ class VAECache:
 
         # Get a list of all the files to process (customize as needed)
         files_to_process = []
-        logger.debug(f'Beginning processing of VAECache directory {directory}')
+        logger.debug(f"Beginning processing of VAECache directory {directory}")
         for subdir, _, files in os.walk(directory):
             for file in files:
-                if file.endswith(('.png', '.jpg', '.jpeg')):
-                    logger.debug(f'Discovered image: {os.path.join(subdir, file)}')
+                if file.endswith((".png", ".jpg", ".jpeg")):
+                    logger.debug(f"Discovered image: {os.path.join(subdir, file)}")
                     files_to_process.append(os.path.join(subdir, file))
 
         # Iterate through the files, displaying a progress bar
-        for filepath in tqdm(files_to_process, desc='Processing images'):
+        for filepath in tqdm(files_to_process, desc="Processing images"):
+            # Create a hash based on the filename
+            file_hash = self.create_hash(filepath)
+            filename = os.path.join(self.cache_dir, file_hash + ".pt")
+
+            # If processed file already exists, skip processing for this image
+            if os.path.exists(filename):
+                logger.debug(
+                    f"Skipping processing for {filepath} as cached file {filename} already exists."
+                )
+                continue
+
             # Open the image using PIL
             try:
-                logger.debug(f'Loading image: {filepath}')
+                logger.debug(f"Loading image: {filepath}")
                 image = Image.open(filepath)
-                image = image.convert('RGB')
+                image = image.convert("RGB")
                 image = self._resize_for_condition_image(image, self.resolution)
-                # Create a hash based on the filename
-                file_hash = self.create_hash(filepath)
             except Exception as e:
-                logger.error(f'Encountered error opening image: {e}')
+                logger.error(f"Encountered error opening image: {e}")
                 os.remove(filepath)
                 continue
 
             # Convert the image to a tensor
             try:
-                pixel_values = transform(image).to(self.accelerator.device, dtype=self.vae.dtype)
+                pixel_values = transform(image).to(
+                    self.accelerator.device, dtype=self.vae.dtype
+                )
             except OSError as e:
-                logger.error(f'Encountered error converting image to tensor: {e}')
+                logger.error(f"Encountered error converting image to tensor: {e}")
                 continue
 
             # Process the image with the VAE
             self.encode_image(pixel_values, file_hash)
 
-            logger.debug(f'Processed image {filepath}')
+            logger.debug(f"Processed image {filepath}")
+
     def _resize_for_condition_image(self, input_image: Image, resolution: int):
         input_image = input_image.convert("RGB")
         W, H = input_image.size
